feat(rtc): add revisioned ICE restart signaling
Why: task-260802032916 requires backend ICE restart updates to preserve the active Pion session while rejecting stale or conflicting signaling.

What: add session-scoped Offer revisions, Answer retry caching, candidate dedupe and limits, recovery grace/deadline lifecycle, HTTP status mappings, shared fixtures, and synchronized RTC contract documentation.

Verify: go vet ./...; go test ./...; go test -race ./internal/rtc ./internal/signaling; npm run gate; npm run tasks:check

Risk: multi-instance signaling remains out of scope; revision state is process-local by design.

Refs: task-260802032916-pion-phase-3-backend-ice-restart

+++ sincromisor-server/sincro-rtc-pion-poc/internal/signaling/offer_registry.go
@@ -68,7 +68,7 @@ type offerEntry struct {
 	requestID string
 	sdpHash   [sha256.Size]byte
 	sessionID string
-	revision  int
+	revision  uint64
 	answer    rtc.Answer
 	err       error
 	state     offerState
